requirements.txt
joblib==0.16.0
neat-python==0.92
numpy==1.19.1
pandas==0.25.3
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
